Add support for the trackpad by tweaking the interrupt transfers.
The trackpad doesn't seem standard, it reports absolute positions while its descriptor tells that positions are relative. To get anything useful from the trackpad tweaking the report descriptor won't be enough, as we need to make the positions relative to the first press position, not any fixed origin, so the interrupt transfer data has to be tweaked.