Build for Windows 7 and later.
The driver could also be built for Windows XP and Vista, but the Windows 10 DDK only targets Win7 and later.

Since this seems to make the WDF CoInstaller required again, re-add it in the .inf file and take advantage of the opportunity to bring back Win XP and Vista support as well.

This fixes issue #2.